pipeline-saude-rj/requirements.txt
# Core ETL and Data Processing
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
pyarrow==12.0.1
fastparquet==2023.7.0

# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-auth==2.22.0
google-auth-oauthlib==1.0.0

# dbt
dbt-core==1.5.2
dbt-bigquery==1.5.3

# Airflow
apache-airflow==2.6.3
apache-airflow-providers-google==10.8.0

# Data Quality
great-expectations==0.17.12

# Utilities
python-dotenv==1.0.0
click==8.1.6
pydantic==2.1.1
sqlalchemy==1.4.49

# Development and Testing
pytest==7.4.0
black==23.7.0
isort==5.12.0
flake8==6.0.0
mypy==1.4.1

# Jupyter for exploration
jupyter==1.0.0
ipykernel==6.25.0

# Logging and Monitoring
structlog==23.1.0